# apps/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Movimento 7 API
  version: 0.1.0
  description: API versionada da plataforma Movimento 7.
servers:
  - url: https://movimento7.com.br/api/v1
    description: Produção via proxy same-origin
...
paths:
...
  /admin/auth/login:
    post:
      tags: [admin]
      operationId: adminLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
  /admin/auth/logout:
    post:
      tags: [admin]
      operationId: adminLogout
      parameters: [{ $ref: '#/components/parameters/CsrfToken' }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /admin/auth/session:
    get:
      tags: [admin]
      operationId: getAdminSession
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Error' }
  /admin/auth/change-password:
    post:
      tags: [admin]
      operationId: changeAdminPassword
      parameters: [{ $ref: '#/components/parameters/CsrfToken' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password: { type: string, format: password }
                new_password:
                  type: string
                  format: password
                  minLength: 12
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
  /admin/auth/password-reset/request:
    post:
      tags: [admin-auth]
      operationId: requestAdminPasswordReset
      summary: Solicita recuperação sem revelar se a conta existe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email, maxLength: 180 }
      responses:
        '202': { $ref: '#/components/responses/Success' }
        '429': { $ref: '#/components/responses/Error' }
  /admin/auth/password-reset/confirm:
    post:
      tags: [admin-auth]
      operationId: confirmAdminPasswordReset
      summary: Redefine a senha com token de uso único
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, new_password, password_confirmation]
              properties:
                token: { type: string, minLength: 20, maxLength: 200 }
                new_password: { type: string, minLength: 12, maxLength: 256 }
                password_confirmation: { type: string, minLength: 12, maxLength: 256 }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '422': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
...
components:
  parameters:
...
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: { type: string }
  schemas:
    Envelope:
      type: object
      required: [data, meta, error, request_id]
      properties:
        data: {}
        meta: { type: object, additionalProperties: true }
        error:
          oneOf: [{ type: 'null' }, { $ref: '#/components/schemas/ApiError' }]
        request_id: { type: [string, 'null'] }
    ApiError:
      type: object
      required: [code, message, fields]
      properties:
        code: { type: string }
        message: { type: string }
        fields:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
...
  responses:
    Success:
      description: Resposta bem-sucedida
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
    Error:
      description: Erro esperado e seguro
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }